Diff Staging: create branches instead of tags
Branches are ephemeral by nature; tags are not. Creating a tag for every
revision of every diff accumulates excessively in repositories which are
high churn. Additionally, reusing a repository as a staging repo,
instead of creating a new one, forces all of these tags to be pulled
down for all developers.

Let's use branches instead of tags, as it is trivial to remove remote,
old branches as they accumulate. Removing tags is non-trivial and likely
to result in a developer accidentally re-uploading all of the deleted
tags.