Add support for Check-Prompt-Resolve scripts in arc diff.
Check-Prompt-Resolve (CPR) scripts are custom checks to be added
before unit and lint tests, which allow us to prompt the user to
run a (potentially heavyweight) fix script if a provided
lightweight check fails.

It adds the following optional entry to `.arcconfig`:

```
"uber.differential.check_prompt_resolve": [
    {
      "check-script" : "./path/to/quick_check.sh",
      "prompt" : "[Explanation]. Fix it? (This can take a few minutes)",
      "default-to" : true,
      "resolve-script" : "./path/to/fix_script.sh"
    }
  ]
```

For each object in the list, `arc diff` will run the check-script
unconditionally. If it succeeds, it will move on to the next CPR or
to lint/unit tests. If the check fails, it will prompt the user with
"prompt" and either [Y/n] (if "default-to" is "true") or [y/N] (if
"default-to" is "false"). If the user answers yes, it will run
the resolve-script.

The reason for this workflow vs `arc lint` autofix, is that `arc lint`
requires the patch to be provided at the same time as the fix. The
use case for CPRs is that of an expensive script needed to fix the
diff, which is rarely run, and only needed if some easy to check
condition is true.