linter: make govet a noop linter
`go tool vet` was finally removed in Go 1.12. The old invocation allowed
files to be individually passed but the new one wants import paths to
packages. This is a major change to how `ArcanistGoVetLinter` needs to
work.

For the time being, disable `govet` linting by turning it into a noop
linter.